+++ Code/Views/BOTActionInputView.m
@@ -32,6 +32,7 @@
 NSUInteger const STPickerRowHeight = 52;
 CGFloat const STBorderWidth = 1.0;
 NSUInteger const STCornerRadius = 6;
+NSUInteger const STButtonOffset = 8;
 
 - (id)initWithSelectionTitles:(NSArray *)titles actions:(NSArray *)actions
 {
@@ -40,45 +41,32 @@ NSUInteger const STCornerRadius = 6;
         self.titles = titles;
         self.actions = actions;
         
-        UIColor *blue = BOTBlueColor();
-        self.layer.borderColor = blue.CGColor;
-        self.layer.borderWidth = STBorderWidth;
-        self.layer.cornerRadius = STCornerRadius;
-        self.clipsToBounds = YES;
-        self.backgroundColor = blue;
-        
         if (titles.count > 0) {
-            _button1 = [BOTActionButton initWithTitle:(NSString *)titles[0]];
+            _button1 = [BOTActionButton initWithTitle:(NSString *)titles[0] verticalOffset:STButtonOffset];
             [_button1 addTarget:self action:@selector(buttonTapped:) forControlEvents:UIControlEventTouchUpInside];
             _button1.tag = 0;
             [self addSubview:_button1];
         }
         
         if (titles.count > 1) {
-            _button2 = [BOTActionButton initWithTitle:(NSString *)titles[1]];
+            _button2 = [BOTActionButton initWithTitle:(NSString *)titles[1] verticalOffset:STButtonOffset];
             [_button2 addTarget:self action:@selector(buttonTapped:) forControlEvents:UIControlEventTouchUpInside];
             _button1.tag = 1;
             [self addSubview:_button2];
         }
         
         if (titles.count > 2) {
-            _button3 = [BOTActionButton initWithTitle:(NSString *)titles[2]];
+            _button3 = [BOTActionButton initWithTitle:(NSString *)titles[2] verticalOffset:STButtonOffset];
             [_button3 addTarget:self action:@selector(buttonTapped:) forControlEvents:UIControlEventTouchUpInside];
             _button3.tag = 2;
             [self addSubview:_button3];
         }
-        
-        if (titles.count > 3) {
-            _button4 = [BOTActionButton initWithTitle:(NSString *)titles[3]];
-            [_button4 addTarget:self action:@selector(buttonTapped:) forControlEvents:UIControlEventTouchUpInside];
-            _button4.tag = 3;
-            [self addSubview:_button4];
-        }
+
         
         [self configureAutoLayoutConstraints];
     }
     return self;
-} 
+}
 
 - (void)buttonTapped:(UIButton *)sender
 {
@@ -96,8 +84,8 @@ NSUInteger const STCornerRadius = 6;
 {
     // Button 1
     if (self.titles.count > 0) {
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button1 attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeLeft multiplier:1.0 constant:0]];
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button1 attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeWidth multiplier:1.0 constant:0]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button1 attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeWidth multiplier:0.6 constant:0.0]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button1 attribute:NSLayoutAttributeCenterX relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeCenterX multiplier:1.0 constant:0]];
         [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button1 attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeTop multiplier:1.0 constant:0]];
         self.btn1HeightConstraint = [NSLayoutConstraint constraintWithItem:self.button1 attribute:NSLayoutAttributeHeight relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeHeight multiplier:0.0f constant:0.0f];
         [self addConstraint:self.btn1HeightConstraint];
@@ -105,8 +93,8 @@ NSUInteger const STCornerRadius = 6;
     
     // Button 2
     if (self.titles.count > 1) {
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button2 attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeLeft multiplier:1.0 constant:0]];
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button2 attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeWidth multiplier:1.0 constant:0]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button2 attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeWidth multiplier:0.6 constant:0.0]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button2 attribute:NSLayoutAttributeCenterX relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeCenterX multiplier:1.0 constant:0]];
         [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button2 attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationEqual toItem:self.button1 attribute:NSLayoutAttributeBottom multiplier:1.0 constant:STBorderWidth]];
         self.btn2HeightConstraint = [NSLayoutConstraint constraintWithItem:self.button2 attribute:NSLayoutAttributeHeight relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeHeight multiplier:0.0f constant:0.0f];
         [self addConstraint:self.btn2HeightConstraint];
@@ -114,30 +102,20 @@ NSUInteger const STCornerRadius = 6;
     
     // Button 3
     if (self.titles.count > 2) {
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button3 attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeLeft multiplier:1.0 constant:0]];
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button3 attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeWidth multiplier:1.0 constant:0]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button3 attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeWidth multiplier:0.6 constant:0.0]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button3 attribute:NSLayoutAttributeCenterX relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeCenterX multiplier:1.0 constant:0]];
         [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button3 attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationEqual toItem:self.button2 attribute:NSLayoutAttributeBottom multiplier:1.0 constant:STBorderWidth]];
         self.btn3HeightConstraint = [NSLayoutConstraint constraintWithItem:self.button3 attribute:NSLayoutAttributeHeight relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeHeight multiplier:0.0f constant:0.0f];
         [self addConstraint:self.btn3HeightConstraint];
-    }
-    
-    // Button 3
-    if (self.titles.count > 3) {
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button4 attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeLeft multiplier:1.0 constant:0]];
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button4 attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeWidth multiplier:1.0 constant:0]];
-        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.button4 attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationEqual toItem:self.button3 attribute:NSLayoutAttributeBottom multiplier:1.0 constant:STBorderWidth]];
-        self.btn4HeightConstraint = [NSLayoutConstraint constraintWithItem:self.button4 attribute:NSLayoutAttributeHeight relatedBy:NSLayoutRelationEqual toItem:self attribute:NSLayoutAttributeHeight multiplier:0.0f constant:0.0f];
-        [self addConstraint:self.btn4HeightConstraint];
     }
 }
 
 - (void)layoutSubviews
 {
-    CGFloat height = self.frame.size.height / self.titles.count;
+    CGFloat height = self.frame.size.height / 3;
     self.btn1HeightConstraint.constant = height;
     self.btn2HeightConstraint.constant = height;
     self.btn3HeightConstraint.constant = height;
-    self.btn4HeightConstraint.constant = height;
 }
 
 @end